Add files via upload
Data was visualised in different types of plots of different scales. The volatility is the most important trait to find out in time series that which model fits the most. We take first order differencing and then second order and so on until we find out which one is stationarity (ability to predict future value due to mean and variance being independent of time) for the given data and according to which we forecast the data.